requirements.txt
asgiref==3.3.1
aspy.refactor-imports==2.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
cached-property==1.5.2
decorator==4.4.2
Django==3.1.5
django-debug-toolbar==3.2.1
django-filebrowser-no-grappelli==3.8.0
django-livereload-server==0.3.2
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
factory-boy==3.2.0
Faker==8.1.1
jedi==0.18.0
jsmin==2.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
